Avoid computing contentFrame when not using overflow: visible (#39325)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39325

This diff extracts the logic to compute the content bounds of a shadow node (to compute its overflow insets) to a separate method. The new method is renamed as `getContentBounds` because layout metrics already have a method called `getContentFrame` that has a different meaning (the content box of the node, which excludes border and padding).

As a nice side-effect, we can now avoid executing this logic altogether if the node doesn't have overflow: visible (because we weren't assigning the result anywhere anyway).

NOTE: This method is made public because we need it to compute `scrollWidth` and `scrollHeight` in a following diff.

Changelog: [internal]

Reviewed By: NickGerleman

Differential Revision: D49020219

fbshipit-source-id: 7a65abf8523cb1dbcf0f92565fbfc228083a7d21