*Attempt* to use the hermesversion in the podfile.lock cache (#39379)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39379

We are seeing some failures in circleci that looks weird. By inspecting the tar of Hermes that should be used, they all have the right files. But then, the artifct actually used is missing something.
By looking at the log of the pod install step, it seems that the hermes-engine is not installed, thus using something stored in a previous cache.

This change should take into consideration the hermesversion also for the podfile.lock, so that we do have different Podfile.lock based on different versions of hermes.
I'm also bumping the versions of the keys to reset the caches and to keep them in sync.

## Changelog:
[Internal] - USe the hermesversion checksum in the podfile.lock keys.

Reviewed By: cipolleschi

Differential Revision: D49134827

fbshipit-source-id: c0e1dbc11ec61825f615315aa6215806b7577845