Unblock `yarn android` on main (#39413)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39413

`yarn android` on main is currently broken due to the CLI attempting to search for the manifest inside the src/androidTest folder.
This fixes it by specifying the exact path of the Android Manifest.
The fix to the CLI is also pending to prevent the CLI from searching inside test folders.

Fix for the CLI is here:
- https://github.com/react-native-community/cli/pull/2075

Changelog:
[Internal] [Changed] - Unblock `yarn android` on main

Reviewed By: huntie

Differential Revision: D49190626

fbshipit-source-id: 99309f17cb08a33be2a565f5faa29130862686ea