cleanup warnOnLegacyNativeModuleSystemUse (#39417)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39417

## Changelog
[Android][Breaking] - Deleting warnOnLegacyNativeModuleSystemUse

this has not fired in over a year, we should clean it up. there's an argument we want to keep this for consumers that are in the middle of upgrading, but personally i think that this needs to be built into the interop layer - currently we can only configure this in Fb specific infra.

Reviewed By: cortinico

Differential Revision: D49209540

fbshipit-source-id: 0e2bb293999ad356e564a0acf069211a7a205f21